requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
astunparse==1.6.3
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.59.2
fpdf==1.7.2
gast==0.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
gTTS==2.5.4
h5py==3.14.0
httplib2==0.31.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
joblib==1.5.2
keras==3.10.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
llvmlite==0.43.0
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.3
mdurl==0.1.2
ml_dtypes==0.5.3
msgpack==1.1.1
mysql-connector-python==9.4.0
namex==0.1.0
numba==0.60.0
numpy==1.26.4
oauthlib==3.3.1
opencv-python==4.8.1.78
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==1.5.3
Pillow==10.0.1
platformdirs==4.4.0
pooch==1.8.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzbar==0.1.9
qrcode==8.2
reportlab==4.4.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
six==1.17.0
soundfile==0.12.1
soxr==1.0.0
SQLAlchemy==2.0.43
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.66.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.2
zipp==3.23.0
